+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - sjasmplus "--lstlab" labels are parsed now.
 
 # 1.3.1
+- Corrected display of memory slots for zsim and cspect.
 - Removed 'replaceAll'.
 
 # 1.3.0